jt/refactor config: config and service recipes
This commit moves the configuration directory and template resources
to a separate config recipe (as it was prior to commit 9c6deece65). It
also moves the service resources to a separate recipe. These changes
allow consumers of the cookbook to be more selective about what they
use from it, or use the default to get the full bowl of awesome.

* Add libraries/helpers methods for platform specific config that use
  the `platform_specific_path`s.
* Move the services to a separate recipe w/ a platform family case
  statement.
* Service resources subscribe to the config file.
* Update the specs.