Default to :latest version of push jobs
If you don't specify a version, it defaults to 'nil'. Until chef 13,
this wasn't a problem, as it would result in the resource's default
value of :latest being used. However, in Chef 13, this results in the
version literally being set to nil, causing an error.

This change simply adds :latest as an explicit fallback option when
setting the version to install.

Signed-off-by: Mark Harrison <mark@mivok.net>